Skip much of aux_config.mk
This was showing up in the top-ten longest makefiles during the product
config and cleanspec kati runs sometimes, even though it's really never
used. So instead of running the sometimes-expensive find command, skip it
unless an AUX OS was specified.

At some point we should remove this entirely, but that's a larger change
that doesn't need to happen yet.

Test: build-aosp_arm.ninja is the same before/after
Change-Id: I0ea10193d2f07e1717224e9020a7defaf6191c05
Signed-off-by: Udit Karode <udit.karode@gmail.com>